Rename build drop to pipeline artifact.
