+++ src/Misc/layoutbin/en-US/strings.json
@@ -38,21 +38,21 @@
   "AutoLogonPoliciesWarningsHeader": "Following policies may affect the autologon:",
   "BuildCommandNotFound": "##vso[build.{0}] is not a recognized command for Build command extension. Please reference documentation (http://go.microsoft.com/fwlink/?LinkId=817296)",
   "BuildDirLastUseTIme": "The last time build directory '{0}' been used is: {1}",
-  "BuildIdIsNotAvailable": "Trying to download build drop in '{0}' environment but build id is not present. Can only download a build drop in '{1}' environment if the artifact is a build.",
+  "BuildIdIsNotAvailable": "Trying to download pipeline artifact in '{0}' environment but build id is not present. Can only download a pipeline artifact in '{1}' environment if the artifact is a build.",
   "BuildIdIsNotValid": "Build Id is not valid: {0}",
   "BuildNumberRequired": "Build number is required.",
   "BuildTagAddFailed": "Build tag '{0}' was not added successfully.",
   "BuildTagRequired": "Build tag is required.",
   "BuildTagsForBuild": "Build '{0}' has following tags now: {1}",
   "CannotChangeParentTimelineRecord": "Can't change parent timeline record of an existing timeline record.",
-  "CannotDownloadFromCurrentEnvironment": "Cannot download a build drop from {0} environment.",
+  "CannotDownloadFromCurrentEnvironment": "Cannot download a pipeline artifact from {0} environment.",
   "CannotFindHostName": "Cannot find VSTS organization name from server url: '{0}'",
   "CanNotFindService": "Cannot find service {0}",
   "CanNotGrantPermission": "Cannot grant LogonAsService permission to the user {0}",
   "CanNotStartService": "Cannot start the service. Check the logs for more details.",
   "CanNotStopService": "Cannot stop the service {0} in a timely fashion.",
   "CannotUploadFile": "Cannot upload file because file location is not specified.",
-  "CannotUploadFromCurrentEnvironment": "Cannot upload to a build drop from {0} environment.",
+  "CannotUploadFromCurrentEnvironment": "Cannot upload to a pipeline artifact from {0} environment.",
   "CannotUploadSummary": "Cannot upload summary file, summary file location is not specified.",
   "ClockSkewStopRetry": "Stopped retrying OAuth token request exception after {0} seconds.",
   "CodeCoverageCommandNotFound": "##vso[codecoverage.{0}] is not a recognized command for Task command extension. Please reference documentation (http://go.microsoft.com/fwlink/?LinkId=817296)",
@@ -569,8 +569,8 @@
   "UpgradeToLatestGit": "To get a better Git experience, upgrade your Git to at least version '{0}'. Your current Git version is '{1}'.",
   "UploadArtifact": "Upload Artifact",
   "UploadArtifactCommandNotSupported": "Uploading server artifact is not supported in {0}.",
-  "UploadArtifactFinished": "Uploading drop finished.",
-  "UploadingBuildDrop": "Uploading drop from {0} for build #{1}",
+  "UploadArtifactFinished": "Uploading pipeline artifact finished.",
+  "UploadingPipelineArtifact": "Uploading pipeline artifact from {0} for build #{1}",
   "UploadToFileContainer": "Upload '{0}' to file container: '{1}'",
   "UserName": "user name",
   "UserShutdownAgent": "The agent has received a shutdown signal. This can happen when the agent service is stopped, or a manually started agent is canceled.",
